Phase 2: move interop.py into dense_evolution.interop/ subpackage
- git mv dense_evolution/interop.py -> dense_evolution/interop/qiskit_pennylane.py
- new interop/__init__.py re-exports the public API, so `from dense_evolution.interop
  import run_qiskit_circuit` (etc.) keeps working unchanged -- no compat shim needed
  since the subpackage name matches the old module name, so the dotted import path
  dense_evolution.interop is unaffected
- fixed the module's two now-broken relative imports (.parser/.simulator -> ..parser/..simulator)
  after the physical move
- updated tests/test_interop*.py to import the private/internal test hooks
  (HAS_QISKIT, HAS_PENNYLANE, HAS_STIM, _require_qiskit, _to_qiskit_bit_order)
  directly from dense_evolution.interop.qiskit_pennylane, since monkeypatching
  those only takes effect against the real defining module
- updated docs/api/interop.md and pyproject.toml packages list accordingly
- also fixes the README's top-level Architecture diagram, stale since both the
  Phase 1 tools/research move (PR #75) and the utils/ subpackage split (PR #79)

Full suite verified clean: 800 passed, 17 skipped, 0 failed.
Part of #76.
