+++ dense_evolution/interop/__init__.py
@@ -1,0 +1,18 @@
+"""Interop subpackage: bridges to Qiskit and PennyLane circuits/backends."""
+from .qiskit_pennylane import (
+    from_qiskit,
+    from_pennylane,
+    run_qiskit_circuit,
+    run_pennylane_circuit,
+    noise_model_from_qiskit_backend,
+    to_stim,
+)
+
+__all__ = [
+    "from_qiskit",
+    "from_pennylane",
+    "run_qiskit_circuit",
+    "run_pennylane_circuit",
+    "noise_model_from_qiskit_backend",
+    "to_stim",
+]
