+++ dense_evolution/interop/qiskit_pennylane.py
@@ -55,8 +55,8 @@ import warnings
 from typing import Optional, Tuple
 import numpy as np
 
-from .parser import QASMParser, QASMCircuit
-from .simulator import DenseSVSimulator
+from ..parser import QASMParser, QASMCircuit
+from ..simulator import DenseSVSimulator
 
 try:
     import qiskit.qasm2 as _qasm2
