+++ docs/api/interop.md
@@ -1,3 +1,3 @@
 # Interop (Qiskit / PennyLane)
 
-::: dense_evolution.interop
+::: dense_evolution.interop.qiskit_pennylane
