+++ tests/test_interop_calibration_noise.py
@@ -22,7 +22,7 @@ import numpy as np
 import pytest
 
 import dense_evolution as de
-from dense_evolution import interop
+from dense_evolution.interop import qiskit_pennylane as interop
 from dense_evolution.interop import noise_model_from_qiskit_backend
 from dense_evolution.registry import NoiseModel
 from dense_evolution.measurement import statevector_fidelity
